created LoginManager to do login jobs seperately from UserManager
