+++ src/main/java/manager/UserManager.java
@@ -10,19 +10,9 @@ import java.util.Objects;
  * A manager that manages all Users.
  */
 public class UserManager extends Manager<User>{
-    private boolean loggedIn = false; // by default, no user logged in
-    private Object currUser = null; // by default, no current user who is logged in
 
     public UserManager() {
         super();
-    }
-
-    public User getCurrUser() throws Exception {
-        if (this.currUser instanceof User) {
-            return (User) this.currUser;
-        } else {
-            throw new Exception("There's no logged-in user.");
-        }
     }
 
     public void createNewUser(String name, String password) {
@@ -32,50 +22,15 @@ public class UserManager extends Manager<User>{
     }
 
     /**
-     * Return the User with given name and password
-     * @param name name of a User
-     * @param password password of a User (matches the name)
-     * @return User or if not found (which can't be the case; included for completeness reason) return null
-     */
-    private Object findUser(String name, String password) {
-        for (User user : this.idToItem.values()) {
-            if (Objects.equals(user.getName(), name) && Objects.equals(user.getPassword(), password)) {
-                return user;
-            }
-        }
-        return null;
-    }
-
-    public void logInUser(String name, String password) throws Exception {
-        if (this.loggedIn) {
-            throw new Exception("Please sign off before logging in.");
-        } else if (this.findUser(name, password) == null) {
-            throw new Exception("Invalid name or password. If you are new, please create an account first.");
-        }
-        this.loggedIn = true;
-        this.currUser = this.findUser(name, password);
-    }
-
-    public void SignOffUser() throws Exception {
-        if (!this.loggedIn) {
-            throw new Exception("Already signed off.");
-        }
-        this.loggedIn = false;
-        this.currUser = null;
-    }
-
-    /**
      * Change name/password of current user.
      * @param func 'N' for name or 'P' for password
      * @param newInfo new name or new password
      */
-    public void changeInfo(char func, String newInfo) {
-        if (loggedIn) {
-            if (func == 'N') {
-                ((User) this.currUser).changeName(newInfo);
-            } else if (func == 'P') {
-                ((User) this.currUser).changePassword(newInfo);
-            }
+    public void changeInfo(User user, char func, String newInfo) {
+        if (func == 'N') {
+            user.changeName(newInfo);
+        } else if (func == 'P') {
+            user.changePassword(newInfo);
         }
     }
